Use current map center as proximity hint
When geocoding pass the current map center as 'proximity', which is a
new-ish feature of the OpenCage API that adjusts the order of results
that come back from the geocoder.

For example if the map is centered on London, UK then a search for
Croydon comes back with CR0 1SZ, but if the map is centered on Sydney,
Australia then a search for Croydon comes back with Croydon NSW 2132.